requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aio-pika==9.5.4
redis[hiredis]==5.2.1
google-generativeai==0.8.4
httpx==0.28.1
pydantic-settings==2.7.1
